+++ examples/vitepress/docs/changelog.md
@@ -72,7 +72,6 @@ Custom object for tracking sales orders.
 ### Speaker__c
 
 Represents a speaker at an event.
-### VisibleCMT__mdt
 
 ## New or Removed Fields to Custom Objects or Standard Objects
 
@@ -81,3 +80,11 @@ These custom fields have been added or removed.
 ### Contact
 
 - New Field: PhotoUrl__c. URL of the contact&#x27;s photo
+
+## New or Removed Custom Metadata Type Records
+
+These custom metadata type records have been added or removed.
+
+### VisibleCMT__mdt
+
+- New Custom Metadata Record: Some_Record_1